requirements.txt
requests
python-dotenv
ta
pandas
schedule
websocket-client
